wip: refactoring moving logic
