--- src/redux/actionTypes.js
@@ -1,3 +1,0 @@
-export const INCREMENT = "INCREMENT";
-export const DECREMENT = "DECREMENT";
-export const RESET = "RESET";
